+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Use `getenv` instead of `$_SERVER` to fetch env vars
 - Set default values for environment variables in functional tests
+- Remove the option to skip tests that depend on the database
 
 ## [1.2.0] - 2020-07-03
 [1.2.0]: https://github.com/Smile-SA/gdpr-dump/compare/1.1.1...1.2.0